requirements.txt
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv
pandas
numpy
tqdm
sklearn
transformers
pytorch-lightning
wandb
tokenizers
scikit-learn
scipy
allennlp
fairseq
fastBPE
nlpaug
textattack
pandas-bokeh
spacy
gensim
matplotlib
pyLDAvis
captum
eli5
torch
sacremoses
subword_nmt
tensorflow_addons
